requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
blessed==1.20.0
blinker==1.7.0
certifi==2023.7.22
cftime==1.6.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
editor==1.6.6
exceptiongroup==1.1.3
executing==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
inquirer==3.2.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.3
jupyter_client==8.3.1
jupyter_core==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.7
netCDF4==1.6.4
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydobot==1.3.2
Pygments==2.16.1
pygmt==0.10.0
pyserial==3.4
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
readchar==4.0.5
rich==13.7.0
runs==1.2.2
shellingham==1.5.4
six==1.16.0
stack-data==0.6.2
termcolor==2.4.0
tinydb==4.8.0
tornado==6.3.3
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
wcwidth==0.2.6
Werkzeug==3.0.1
xarray==2023.9.0
xmod==1.8.1
yaspin @ https://github.com/pavdmyt/yaspin/archive/master.zip
